First puzzle day 17

+++ src/test/java/com/pgoellner/advent22/day17/LavaTetrisTest.java
@@ -2,7 +2,6 @@ package com.pgoellner.advent22.day17;
 
 import com.pgoellner.advent22.PuzzleInput;
 import com.pgoellner.advent22.TestInput;
-import com.pgoellner.advent22.day16.PressureCalculator;
 
 import java.util.List;
 
@@ -14,10 +13,16 @@ public class LavaTetrisTest {
 
     public static void main(String[] args) {
         tower_height_of_3068();
+//        tower_height_of_1514285714288();
     }
 
     private static void tower_height_of_3068() {
-        int towerHeight = new LavaTetris().simulateStoneFall(puzzleInput);
+        long towerHeight = new LavaTetris().simulateStoneFall(puzzleInput);
         assert towerHeight == 3068 : "Expected tower height of 3068, but got " + towerHeight;
     }
+
+//    private static void tower_height_of_1514285714288() {
+//        long towerHeight = new LavaTetris().simulateStoneFall(puzzleInput, 1_000_000_000_000L);
+//        assert towerHeight == 1_514_285_714_288L : "Expected tower height of 1514285714288, but got " + towerHeight;
+//    }
 }
